requirements.txt
pandas==2.2.2
requests==2.32.3
pandas-gbq==0.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
pyarrow==17.0.0